Make a Python module
one problem with the current design of the webapp is that each thread
gets its own ana process, each of which uses a few megs of RAM.  It may
be that using an ana module is preferable.

One shortcoming, at least initially, is that this design will not
incrementally send the solutions to the user, as it's not implemented
as a generator.