Fix "must contain" with too many of a letter
If the "must contain" (rs) set had too many uses of a character,
the initial letter counts (ww) could be nonsense.

Now, if you do a bogus search like `munmunii -- iii` (commandline)
or `munmun =iii` (web), you'll get 0 results. (it'll somewhat
confusingly report that it reached a max of 0 searches, but so
be it)

Closes #1